+++ unstoppable-ios-app/domains-manager-ios/Services/MessagingService/MessagingService.swift
@@ -9,20 +9,20 @@ import Foundation
 
 final class MessagingService {
 
-    private let apiService: MessagingAPIServiceProtocol
-    private let channelsApiService: MessagingChannelsAPIServiceProtocol
-    private let webSocketsService: MessagingWebSocketsServiceProtocol
-    private let channelsWebSocketsService: MessagingChannelsWebSocketsServiceProtocol
-    private let storageService: MessagingStorageServiceProtocol
-    private let decrypterService: MessagingContentDecrypterService
-    private let filesService: MessagingFilesServiceProtocol
-    private let unreadCountingService: MessagingUnreadCountingServiceProtocol
+    let apiService: MessagingAPIServiceProtocol
+    let channelsApiService: MessagingChannelsAPIServiceProtocol
+    let webSocketsService: MessagingWebSocketsServiceProtocol
+    let channelsWebSocketsService: MessagingChannelsWebSocketsServiceProtocol
+    let storageService: MessagingStorageServiceProtocol
+    let decrypterService: MessagingContentDecrypterService
+    let filesService: MessagingFilesServiceProtocol
+    let unreadCountingService: MessagingUnreadCountingServiceProtocol
     
-    private let dataRefreshManager = MessagingServiceDataRefreshManager()
-    private var listenerHolders: [MessagingListenerHolder] = []
-    private var currentUser: MessagingChatUserProfileDisplayInfo?
+    let dataRefreshManager = MessagingServiceDataRefreshManager()
+    private(set) var listenerHolders: [MessagingListenerHolder] = []
+    private(set) var currentUser: MessagingChatUserProfileDisplayInfo?
     
-    private var stateHolder = StateHolder()
+    private(set) var stateHolder = StateHolder()
 
     init(apiService: MessagingAPIServiceProtocol,
          channelsApiService: MessagingChannelsAPIServiceProtocol,
@@ -587,248 +587,7 @@ private extension MessagingService {
     }
 }
 
-// MARK: - Chats
-private extension MessagingService {
-    func refreshChatsForProfile(_ profile: MessagingChatUserProfile, shouldRefreshUserInfo: Bool) {
-        Task {
-            dataRefreshManager.startUpdatingChats(for: profile.displayInfo)
-            var startTime = Date()
-            do {
-                let allLocalChats = try await storageService.getChatsFor(profile: profile)
-                let localChats = allLocalChats.filter { $0.displayInfo.isApproved}
-                let localRequests = allLocalChats.filter { !$0.displayInfo.isApproved}
-                
-                async let remoteChatsTask = updatedLocalChats(localChats, forProfile: profile, isRequests: false)
-                async let remoteRequestsTask = updatedLocalChats(localRequests, forProfile: profile, isRequests: true)
-                
-                let (remoteChats, remoteRequests) = await (remoteChatsTask, remoteRequestsTask)
-                let allRemoteChats = remoteChats + remoteRequests
-                
-                let updatedChats = await refreshChatsMetadata(remoteChats: allRemoteChats,
-                                                              localChats: allLocalChats,
-                                                              for: profile)
-                await storageService.saveChats(updatedChats)
-                
-                let updatedStoredChats = try await storageService.getChatsFor(profile: profile)
-                let chatsDisplayInfo = updatedStoredChats.sortedByLastMessage().map({ $0.displayInfo })
-                notifyListenersChangedDataType(.chats(chatsDisplayInfo, profile: profile.displayInfo))
-                Debugger.printTimeSensitiveInfo(topic: .Messaging,
-                                                "to refresh chats list for \(profile.wallet)",
-                                                startDate: startTime,
-                                                timeout: 3)
-                
-                if shouldRefreshUserInfo {
-                    startTime = Date()
-                    await refreshUsersInfoFor(profile: profile)
-                    Debugger.printTimeSensitiveInfo(topic: .Messaging,
-                                                    "to refresh users info for chats list for \(profile.wallet)",
-                                                    startDate: startTime,
-                                                    timeout: 3)
-                }
-                dataRefreshManager.stopUpdatingChats(for: profile.displayInfo)
-            } catch {
-                Debugger.printFailure("Failed to refresh chats list for \(profile.wallet) with error: \(error.localizedDescription)")
-            }
-        }
-    }
-    
-    func updatedLocalChats(_ localChats: [MessagingChat],
-                           forProfile profile: MessagingChatUserProfile,
-                           isRequests: Bool) async -> [MessagingChat] {
-        var remoteChats = [MessagingChat]()
-        let limit = 30
-        var page = 1
-        while true {
-            do {
-                let chatsPage: [MessagingChat]
-                if isRequests {
-                    chatsPage = try await apiService.getChatRequestsForUser(profile, page: 1, limit: limit)
-                } else {
-                    chatsPage = try await apiService.getChatsListForUser(profile, page: 1, limit: limit)
-                }
-                
-                remoteChats.append(contentsOf: chatsPage)
-                if !apiService.capabilities.isSupportChatsListPagination || chatsPage.count < limit {
-                    /// Loaded all chats
-                    break
-                } else if let lastPageChat = chatsPage.last,
-                          let localChat = localChats.first(where: { $0.displayInfo.id == lastPageChat.displayInfo.id }),
-                          lastPageChat.isUpToDateWith(otherChat: localChat) {
-                    /// No changes for other chats
-                    break
-                } else {
-                    page += 1
-                }
-            } catch {
-                break
-            }
-        }
-        
-        await storageService.saveChats(remoteChats)
-        
-        return remoteChats
-    }
-    
-    func refreshChatsMetadata(remoteChats: [MessagingChat],
-                              localChats: [MessagingChat],
-                              for profile: MessagingChatUserProfile) async -> [MessagingChat] {
-        var updatedChats = [MessagingChat]()
-        
-        await withTaskGroup(of: MessagingChat.self, body: { group in
-            for remoteChat in remoteChats {
-                group.addTask {
-                    if !self.apiService.capabilities.isRequiredToReloadLastMessage,
-                       let localChat = localChats.first(where: { $0.displayInfo.id == remoteChat.displayInfo.id }),
-                       localChat.isUpToDateWith(otherChat: remoteChat) {
-                        return localChat
-                    } else {
-                        if var lastMessage = try? await self.apiService.getMessagesForChat(remoteChat,
-                                                                                           before: nil,
-                                                                                           cachedMessages: [],
-                                                                                           fetchLimit: 1,
-                                                                                           isRead: false,
-                                                                                           for: profile,
-                                                                                           filesService: self.filesService).first {
-                            
-                            var updatedChat = remoteChat
-                            if let storedMessage = await self.storageService.getMessageWith(id: lastMessage.displayInfo.id,
-                                                                                            in: remoteChat) {
-                                lastMessage.displayInfo.isRead = storedMessage.displayInfo.isRead
-                            } else {
-                                switch lastMessage.displayInfo.senderType {
-                                case .thisUser:
-                                    lastMessage.displayInfo.isRead = true
-                                case .otherUser:
-                                    lastMessage.displayInfo.isRead = localChats.isEmpty // If loading channels for the first time - messages is read by default.
-                                }
-                            }
-                            
-                            if !lastMessage.displayInfo.senderType.isThisUser && !lastMessage.displayInfo.isRead {
-                                updatedChat.displayInfo.unreadMessagesCount += 1
-                            }
-                            updatedChat.displayInfo.lastMessage = lastMessage.displayInfo
-                            await self.storageService.saveMessages([lastMessage])
-                            return updatedChat
-                        } else {
-                            return remoteChat
-                        }
-                    }
-                }
-            }
-            
-            for await chat in group {
-                updatedChats.append(chat)
-            }
-        })
-        
-        return updatedChats
-    }
-    
-    func refreshUsersInfoFor(profile: MessagingChatUserProfile) async{
-        do {
-            let chats = try await storageService.getChatsFor(profile: profile)
-            await withTaskGroup(of: Void.self, body: { group in
-                for chat in chats {
-                    group.addTask {
-                        if let otherUserInfo = try? await self.loadUserInfoFor(chat: chat) {
-                            for info in otherUserInfo {
-                                await self.storageService.saveMessagingUserInfo(info)
-                            }
-                        }
-                        return Void()
-                    }
-                }
-                
-                for await _ in group {
-                    Void()
-                }
-            })
-            
-            let updatedChats = try await storageService.getChatsFor(profile: profile)
-            notifyListenersChangedDataType(.chats(updatedChats.map { $0.displayInfo }, profile: profile.displayInfo))
-        } catch { }
-    }
-    
-    func loadUserInfoFor(chat: MessagingChat) async throws -> [MessagingChatUserDisplayInfo] {
-        switch chat.displayInfo.type {
-        case .private(let details):
-            let wallet = details.otherUser.wallet
-            if let userInfo = await loadUserInfoFor(wallet: wallet) {
-                return [userInfo]
-            }
-            return []
-        case .group(let details):
-            var infos: [MessagingChatUserDisplayInfo] = []
-            let members = details.allMembers.prefix(3) // Only first 3 members will be displayed on the UI
-            for member in members {
-                if let userInfo = await loadUserInfoFor(wallet: member.wallet) {
-                    infos.append(userInfo)
-                }
-            }
-            return infos
-        }
-    }
-    
-    func loadUserInfoFor(wallet: String) async -> MessagingChatUserDisplayInfo? {
-        if let domain = try? await appContext.udWalletsService.reverseResolutionDomainName(for: wallet.normalized),
-           !domain.isEmpty {
-            let pfpInfo = await appContext.udDomainsService.loadPFP(for: domain)
-            var pfpURL: URL?
-            if let urlString = pfpInfo?.pfpURL,
-               let url = URL(string: urlString) {
-                pfpURL = url
-            }
-            return MessagingChatUserDisplayInfo(wallet: wallet,
-                                                domainName: domain,
-                                                pfpURL: pfpURL)
-        } else if var userInfo = await loadGlobalUserInfoFor(value: wallet) {
-            userInfo.wallet = wallet // Fix lower/uppercase inconsistency issue
-            return userInfo
-        }
-        
-        return nil
-    }
-    
-    // Value can be either wallet address or domain name
-    func loadGlobalUserInfoFor(value: String) async -> MessagingChatUserDisplayInfo? {
-        if let rrInfo = try? await NetworkService().fetchGlobalReverseResolution(for: value.lowercased()) {
-            return MessagingChatUserDisplayInfo(wallet: rrInfo.address,
-                                                domainName: rrInfo.name,
-                                                pfpURL: rrInfo.pfpURLToUse)
-        }
-        return nil
-    }
-    
-    func isNewMessagesAvailable(for profile: MessagingChatUserProfile) async throws -> Bool {
-        if try await isNewMessagesFromAcceptedChatsAvailable(for: profile) {
-            return true
-        }
-        
-        return try await isNewMessagesFromRequestChatsAvailable(for: profile)
-    }
-    
-    func isNewMessagesFromAcceptedChatsAvailable(for profile: MessagingChatUserProfile) async throws -> Bool {
-        let chats = try await apiService.getChatsListForUser(profile, page: 1, limit: 1)
-        
-        return try await isNewMessagesFromChatsAvailable(chats, for: profile)
-    }
-    
-    func isNewMessagesFromRequestChatsAvailable(for profile: MessagingChatUserProfile) async throws -> Bool {
-        let chats = try await apiService.getChatRequestsForUser(profile, page: 1, limit: 1)
-        
-        return try await isNewMessagesFromChatsAvailable(chats, for: profile)
-    }
-    
-    func isNewMessagesFromChatsAvailable(_ chats: [MessagingChat], for profile: MessagingChatUserProfile) async throws -> Bool {
-        guard let latestChat = chats.first else { return false } /// No messages if no chats
-        guard let localChat = await storageService.getChatWith(id: latestChat.displayInfo.id,
-                                                               of: latestChat.userId) else { return true } /// New chat => new message
-        
-        return !localChat.isUpToDateWith(otherChat: latestChat)
-    }
-}
-
-// MARK: - Messages
+// MARK: - Private methods
 private extension MessagingService {
     func sendMessageToBEAsync(message: MessagingChatMessage,
                               messageType: MessagingChatMessageDisplayType,
@@ -856,333 +615,12 @@ private extension MessagingService {
         }
     }
     
-    func notifyReadStatusUpdatedFor(message: MessagingChatMessageDisplayInfo) {
-        let number = unreadCountingService.getNumberOfUnreadMessagesIn(chatId: message.chatId, userId: message.userId)
-        notifyListenersChangedDataType(.messageReadStatusUpdated(message, numberOfUnreadMessagesInSameChat: number))
-    }
-}
-
-// MARK: - Channels
-private extension MessagingService {
-    func refreshChannelsForProfile(_ profile: MessagingChatUserProfile) {
-        Task {            
-            dataRefreshManager.startUpdatingChannels(for: profile.displayInfo)
-            let startTime = Date()
-            do {
-                let storedChannels = try await storageService.getChannelsFor(profile: profile)
-                
-                async let channelsTask = Utilities.catchingFailureAsyncTask(asyncCatching: {
-                    try await channelsApiService.getSubscribedChannelsForUser(profile)
-                }, defaultValue: [])
-                async let spamChannelsTask = Utilities.catchingFailureAsyncTask(asyncCatching: {
-                    try await channelsApiService.getSpamChannelsForUser(profile)
-                }, defaultValue: [])
-                
-                let (channels, spamChannels) = await (channelsTask, spamChannelsTask)
-                let channelsIds = Set(channels.map { $0.id })
-                let allChannels = channels + spamChannels.filter { !channelsIds.contains($0.id) }
-                
-                let updatedChats = await refreshChannelsMetadata(allChannels, storedChannels: storedChannels).sortedByLastMessage()
-                
-                await storageService.saveChannels(updatedChats, for: profile)
-                
-                let updatedStoredChannels = try await storageService.getChannelsFor(profile: profile)
-                notifyListenersChangedDataType(.channels(updatedStoredChannels, profile: profile.displayInfo))
-                Debugger.printTimeSensitiveInfo(topic: .Messaging,
-                                                "to refresh channels list for \(profile.wallet)",
-                                                startDate: startTime,
-                                                timeout: 3)
-            } catch {
-                Debugger.printFailure("Did fail to refresh channels list for \(profile.wallet)")
-            }
-            dataRefreshManager.stopUpdatingChannels(for: profile.displayInfo)
-        }
-    }
-    
-    func refreshChannelsMetadata(_ channels: [MessagingNewsChannel],
-                                 storedChannels: [MessagingNewsChannel]) async -> [MessagingNewsChannel] {
-        var updatedChannels = [MessagingNewsChannel]()
-        
-        await withTaskGroup(of: MessagingNewsChannel.self, body: { group in
-            for channel in channels {
-                group.addTask {
-                    if var lastMessage = try? await self.channelsApiService.getFeedFor(channel: channel,
-                                                                                       page: 1,
-                                                                                       limit: 1,
-                                                                                       isRead: false).first {
-                        var updatedChannel = channel
-                        if let storedChannel = storedChannels.first(where: { $0.id == channel.id }),
-                           let storedLastMessage = storedChannel.lastMessage,
-                           storedLastMessage.id == lastMessage.id {
-                            lastMessage.isRead = storedLastMessage.isRead
-                        } else {
-                            lastMessage.isRead = true
-                            await self.storageService.saveChannelsFeed([lastMessage],
-                                                                       in: channel)
-                        }
-                        updatedChannel.lastMessage = lastMessage
-                        return updatedChannel
-                    } else {
-                        return channel
-                    }
-                }
-            }
-            
-            for await channel in group {
-                updatedChannels.append(channel)
-            }
-        })
-        
-        return updatedChannels
-    }
-}
-
-// MARK: - Private methods
-private extension MessagingService {
-    func getReverseResolutionDomainItem(for wallet: String) async throws -> DomainItem {
-        let walletsForMessaging = await fetchWalletsAvailableForMessaging()
-        guard let wallet = walletsForMessaging.first(where: { $0.address == wallet.normalized }),
-              let domainInfo = wallet.reverseResolutionDomain else {
-            throw MessagingServiceError.noRRDomainForProfile
-        }
-        return try await appContext.dataAggregatorService.getDomainWith(name: domainInfo.name)
-    }
-    
-    func getDomainEthWalletAddress(_ domain: DomainDisplayInfo) throws -> String {
-        guard let ethAddress = domain.getETHAddress() else { throw MessagingServiceError.domainWithoutWallet }
-        return ethAddress
-    }
-    
-    func getMessagingChatFor(displayInfo: MessagingChatDisplayInfo,
-                             userId: String) async throws -> MessagingChat {
-        try await getMessagingChatWith(chatId: displayInfo.id, userId: userId)
-    }
-    
-    func getMessagingChatWith(chatId: String,
-                              userId: String) async throws -> MessagingChat {
-        guard let chat = await storageService.getChatWith(id: chatId,
-                                                          of: userId) else { throw MessagingServiceError.chatNotFound }
-        
-        return chat
-    }
-  
-    func getUserProfileWith(wallet: String) async throws -> MessagingChatUserProfile {
-        let rrDomain = try await getReverseResolutionDomainItem(for: wallet)
-        return try storageService.getUserProfileFor(domain: rrDomain,
-                                                    serviceIdentifier: apiService.serviceIdentifier)
-    }
-    
-    func replaceCacheMessageAndNotify(_ messageToReplace: MessagingChatMessage,
-                                      with newMessage: MessagingChatMessage) {
-        Task {
-            try? await storageService.replaceMessage(messageToReplace, with: newMessage)
-            notifyListenersChangedDataType(.messageUpdated(messageToReplace.displayInfo, newMessage: newMessage.displayInfo))
-        }
-    }
-    
-    func setupSocketConnection(profile: MessagingChatUserProfile) {
-        Task {
-            do {
-                webSocketsService.disconnectAll()
-                try webSocketsService.subscribeFor(profile: profile,
-                                                   eventCallback: { [weak self] event in
-                    self?.handleWebSocketEvent(event)
-                })
-                
-                channelsWebSocketsService.disconnectAll()
-                try channelsWebSocketsService.subscribeFor(profile: profile,
-                                                   eventCallback: { [weak self] event in
-                    self?.handleWebSocketEvent(event)
-                })
-            }
-        }
-    }
-    
-    func handleWebSocketEvent(_ event: MessagingWebSocketEvent) {
-        Task {
-            func addNewChatMessages(_ chatMessages: [GroupChatMessageWithProfile]) async {
-                guard !chatMessages.isEmpty else { return }
-                
-                if await stateHolder.isSendingMessage,
-                   chatMessages.first(where: { $0.message.displayInfo.senderType.isThisUser }) != nil {
-                    return
-                }
-                
-                await storageService.saveMessages(chatMessages.map({ $0.message }))
-                for messageWithProfile in chatMessages {
-                    let message = messageWithProfile.message
-                    let profile = messageWithProfile.profile
-                    let chatId = message.displayInfo.chatId
-                    
-                    notifyListenersChangedDataType(.messagesAdded([message.displayInfo],
-                                                                  chatId: chatId,
-                                                                  userId: profile.id))
-                    try? await setLastMessageAndNotify(lastMessage: message.displayInfo)
-                }
-            }
-            
-            func addNewChannelFeed(_ feed: MessagingNewsChannelFeed, to channel: MessagingNewsChannel) async throws {
-                var channel = channel
-                let profile = try storageService.getUserProfileWith(userId: channel.userId,
-                                                                    serviceIdentifier: apiService.serviceIdentifier)
-                await storageService.saveChannelsFeed([feed], in: channel)
-                channel.lastMessage = feed
-                await storageService.saveChannels([channel], for: profile)
-                notifyListenersChangedDataType(.channelFeedAdded(feed, channelId: channel.id))
-                notifyChannelsChanged(userId: profile.id)
-            }
-            
-            do {
-                switch event {
-                case .channelNewFeed(let feed, let channelAddress, _):
-                    if let channel = try await getOrFetchChannelOfCurrentUserWithAddress(channelAddress, isCurrentUserSubscribed: true) {
-                        try await addNewChannelFeed(feed, to: channel)
-                    }
-                case .channelSpamFeed(let feed, let channelAddress, _):
-                    if let channel = try await getOrFetchChannelOfCurrentUserWithAddress(channelAddress, isCurrentUserSubscribed: false) {
-                        try await addNewChannelFeed(feed, to: channel)
-                    }
-                case .groupChatReceivedMessage(let message):
-                    let chatMessages = try await convertMessagingWebSocketGroupMessageEntityToMessage(message)
-                    await addNewChatMessages(chatMessages)
-                case .chatReceivedMessage(let message):
-                    let chatMessages = try await convertMessagingWebSocketMessageEntityToMessage(message)
-                    await addNewChatMessages(chatMessages)
-                case .newChat(let webSocketsChat):
-                    let profile = try storageService.getUserProfileWith(userId: webSocketsChat.userId,
-                                                                        serviceIdentifier: webSocketsChat.serviceIdentifier)
-                    guard let chat = webSocketsChat.transformToChatBlock(webSocketsChat, profile) else { return }
-                    
-                    let updatedChats = await refreshChatsMetadata(remoteChats: [chat], localChats: [], for: profile)
-                    await storageService.saveChats(updatedChats)
-                    notifyChatsChanged(wallet: profile.wallet)
-                    await refreshUsersInfoFor(profile: profile)
-                }
-            } catch { }
-        }
-    }
-    
-    func getOrFetchChannelOfCurrentUserWithAddress(_ channelAddress: String,
-                                                   isCurrentUserSubscribed: Bool) async throws -> MessagingNewsChannel? {
-        guard let currentUser else { return nil }
-        
-        let profile = try storageService.getUserProfileWith(userId: currentUser.id,
-                                                            serviceIdentifier: apiService.serviceIdentifier)
-        let cachedChannels = try await storageService.getChannelsFor(profile: profile)
-        if let cachedChannel = cachedChannels.first(where: { $0.channel == channelAddress }) {
-            return cachedChannel
-        }
-        
-        let channels = try await channelsApiService.getChannelsWithIds([channelAddress],
-                                                                       isCurrentUserSubscribed: isCurrentUserSubscribed,
-                                                                       user: profile)
-        return channels.first
-    }
-    
-    func refreshChatsInSameDomain(as chatId: String, userId: String) {
-        Task {
-            do {
-                let chat = try await getMessagingChatWith(chatId: chatId, userId: userId)
-                let profile = try await getUserProfileWith(wallet: chat.displayInfo.thisUserDetails.wallet)
-                refreshChatsForProfile(profile, shouldRefreshUserInfo: false)
-            } catch { }
-        }
-    }
-    
-    func notifyChannelsChanged(userId: String) {
-        Task {
-            do {
-                let profile = try storageService.getUserProfileWith(userId: userId,
-                                                                    serviceIdentifier: apiService.serviceIdentifier)
-                let channels = try await storageService.getChannelsFor(profile: profile)
-                notifyListenersChangedDataType(.channels(channels, profile: profile.displayInfo))
-            } catch { }
-        }
-    }
-    
-    func notifyChatsChanged(wallet: String) {
-        Task {
-            do {
-                let profile = try await getUserProfileWith(wallet: wallet)
-                let chats = try await storageService.getChatsFor(profile: profile)
-                let displayInfo = chats.map { $0.displayInfo }
-                notifyListenersChangedDataType(.chats(displayInfo, profile: profile.displayInfo))
-            } catch { }
-        }
-    }
-    
-    func notifyListenersChangedDataType(_ messagingDataType: MessagingDataType) {
-        Debugger.printInfo(topic: .Messaging, "Will notify listeners about data type: \(messagingDataType.debugDescription)")
-        listenerHolders.forEach { holder in
-            holder.listener?.messagingDataTypeDidUpdated(messagingDataType)
-        }
-    }
-    
-    func setLastMessageAndNotify(lastMessage: MessagingChatMessageDisplayInfo) async throws {
-        guard let chat = await storageService.getChatWith(id: lastMessage.chatId, of: lastMessage.userId) else { return }
-        try await setLastMessageAndNotify(lastMessage, to: chat)
-    }
-    
-    func setLastMessageAndNotify(_ lastMessage: MessagingChatMessageDisplayInfo,
-                                 to chat: MessagingChat) async throws {
-        var updatedChat = chat
-        updatedChat.displayInfo.lastMessage = lastMessage
-        updatedChat.displayInfo.lastMessageTime = lastMessage.time
-        try await storageService.replaceChat(chat, with: updatedChat)
-        notifyChatsChanged(wallet: chat.displayInfo.thisUserDetails.wallet)
-        notifyReadStatusUpdatedFor(message: lastMessage)
-    }
-    
-    func convertMessagingWebSocketMessageEntityToMessage(_ messageEntity: MessagingWebSocketMessageEntity) async throws -> [GroupChatMessageWithProfile] {
-        var messages: [GroupChatMessageWithProfile] = []
-        
-        func getMessageFor(wallet: String, otherUserWallet: String) async throws -> GroupChatMessageWithProfile {
-            let profile = try await getUserProfileWith(wallet: wallet)
-            let chats = try await storageService.getChatsFor(profile: profile)
-            guard let chat = chats.first(where: { $0.displayInfo.type.otherUserDisplayInfo?.wallet == otherUserWallet }) else { throw MessagingServiceError.chatNotFound }
-            guard let message = await messageEntity.transformToMessageBlock(messageEntity, chat, filesService) else { throw MessagingServiceError.failedToConvertWebsocketMessage }
-            return GroupChatMessageWithProfile(message: message, profile: profile)
-        }
-        
-        if let message = try? await getMessageFor(wallet: messageEntity.receiverWallet,
-                                       otherUserWallet: messageEntity.senderWallet) {
-            messages.append(message)
-        }
-        
-        if let message = try? await getMessageFor(wallet: messageEntity.senderWallet,
-                                                  otherUserWallet: messageEntity.receiverWallet) {
-            messages.append(message)
-        }
-    
-        return messages
-    }
-    
-    func convertMessagingWebSocketGroupMessageEntityToMessage(_ messageEntity: MessagingWebSocketGroupMessageEntity) async throws -> [GroupChatMessageWithProfile] {
-        var messages: [GroupChatMessageWithProfile] = []
-        
-        let profiles = try storageService.getAllUserProfiles()
-        for profile in profiles {
-            if let chat = await storageService.getChatWith(id: messageEntity.chatId,
-                                                           of: profile.id),
-            let message = messageEntity.transformToMessageBlock(messageEntity, chat, filesService) {
-                messages.append(.init(message: message, profile: profile))
-            }
-        }
-        
-        return messages
-    }
-    
     func setSceneActivationListener() {
         Task { @MainActor in
             SceneDelegate.shared?.addListener(self)
         }
     }
-    
-    struct GroupChatMessageWithProfile {
-        let message: MessagingChatMessage
-        let profile: MessagingChatUserProfile
-    }
-    
+ 
     func totalUnreadMessagesCountUpdated(_ havingUnreadMessages: Bool) {
         notifyListenersChangedDataType(.totalUnreadMessagesCountUpdated(havingUnreadMessages))
     }
@@ -1213,8 +651,8 @@ extension MessagingService {
     }
 }
 
-// MARK: - Private methods
-private extension MessagingService {
+// MARK: - StateHolder
+extension MessagingService {
     actor StateHolder {
         
         private var sendingMessagesCounter: Int = 0
